server/requirements.txt
websockets
openai-whisper
